+++ scripts/build-ipa.sh
@@ -53,6 +53,7 @@ SWIFT_FILES=(
   "$SRC_DIR/Views/Components/ToastView.swift"
   "$SRC_DIR/Views/Components/RingProgress.swift"
   "$SRC_DIR/Views/Components/AppPalette.swift"
+  "$SRC_DIR/Views/Components/LivePhotoView.swift"
 )
 
 SDK_PATH="$(xcrun --sdk "$SDK_NAME" --show-sdk-path)"